benfica/requirements.txt
requests
bs4
python-telegram-bot
todoist-api-python
Deprecated